+ refactor Enum
