+++ src/main/java/ru/sber/projectATM/core/accounting/Balance.java
@@ -1,4 +1,4 @@
-package ru.sber.projectATM.core.helperСlass.accounting;
+package ru.sber.projectATM.core.accounting;
 
 import lombok.Getter;
 import lombok.Setter;
