+++ src/main/java/ru/sber/projectATM/core/accounting/Currency.java
@@ -1,0 +1,7 @@
+package ru.sber.projectATM.core.accounting;
+
+public enum Currency {
+    RUR,
+    USD,
+    EUR
+}
