+++ src/main/java/ru/sber/projectATM/core/accounting/Status.java
@@ -1,4 +1,4 @@
-package ru.sber.projectATM.core.helperСlass.accounting;
+package ru.sber.projectATM.core.accounting;
 
 public enum Status {
     SUCCESS,
